seperated schedule
