+++ lib/view/networkPage.dart
@@ -1,12 +1,9 @@
-import 'package:Runbhumi/services/EventService.dart';
 import 'package:Runbhumi/services/chatroomServices.dart';
 import 'package:Runbhumi/utils/Constants.dart';
 import 'package:bubble_tab_indicator/bubble_tab_indicator.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_icons/flutter_icons.dart';
-import 'package:intl/intl.dart';
 import '../widget/widgets.dart';
 import 'conversation.dart';
 
@@ -43,16 +40,16 @@ class _NetworkState extends State<Network> {
           body: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
-              Padding(
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 24.0, vertical: 8.0),
-                child: Text(
-                  'Schedule',
-                  style: Theme.of(context).textTheme.headline6,
-                ),
-              ),
-              //schedule
-              Schedule(), //scheduleList: scheduleList)
+              // Padding(
+              //   padding:
+              //       const EdgeInsets.symmetric(horizontal: 24.0, vertical: 8.0),
+              //   child: Text(
+              //     'Schedule',
+              //     style: Theme.of(context).textTheme.headline6,
+              //   ),
+              // ),
+              // //schedule
+              // Schedule(),
               Padding(
                 padding: const EdgeInsets.symmetric(
                   horizontal: 24.0,
@@ -65,16 +62,17 @@ class _NetworkState extends State<Network> {
                 ),
               ),
               ChatsTabs(),
-              // TODO: replace placeholders with actual UI
               Expanded(
                 child: TabBarView(
                   children: [
+                    //direct chats
                     DirectChats(),
+                    //teams chat
                     PlaceholderWidget(),
+                    // b/w teams chat
                     PlaceholderWidget(),
                   ],
                 ),
-                // child: DirectChats(),
               ),
             ],
           ),
@@ -133,13 +131,14 @@ class _DirectChatsState extends State<DirectChats> {
                         onTap: () {
                           //Sending the user to the chat room
                           Navigator.push(
-                              context,
-                              MaterialPageRoute(
-                                  builder: (context) => Conversation(
-                                        chatRoomId: asyncSnapshot
-                                            .data.documents[index]
-                                            .get('chatRoomId'),
-                                      )));
+                            context,
+                            MaterialPageRoute(
+                              builder: (context) => Conversation(
+                                chatRoomId: asyncSnapshot.data.documents[index]
+                                    .get('chatRoomId'),
+                              ),
+                            ),
+                          );
                         },
                         leading: Icon(Icons.person),
                         trailing: Icon(Icons.send),
@@ -397,212 +396,7 @@ class UserSearchDirect extends SearchDelegate<ListView> {
   }
 }
 
-class Schedule extends StatefulWidget {
-  // const Schedule({
-  //   Key key,
-  //   @required this.scheduleList,
-  // }) : super(key: key);
 
-  // final List scheduleList;
-
-  @override
-  _ScheduleState createState() => _ScheduleState();
-}
-
-class _ScheduleState extends State<Schedule> {
-  Stream currentFeed;
-  void initState() {
-    super.initState();
-    Firebase.initializeApp().whenComplete(() {
-      print("completed");
-      setState(() {});
-    });
-    getUserInfoEvents();
-  }
-
-  getUserInfoEvents() async {
-    EventService().getCurrentUserFeed().then((snapshots) {
-      setState(() {
-        currentFeed = snapshots;
-        print("we got the data + ${currentFeed.toString()} ");
-      });
-    });
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-  }
-
-  Widget feed() {
-    return StreamBuilder(
-      stream: currentFeed,
-      builder: (context, asyncSnapshot) {
-        print("Working");
-        return asyncSnapshot.hasData
-            ? asyncSnapshot.data.documents.length > 0
-                ? ListView.builder(
-                    itemCount: asyncSnapshot.data.documents.length,
-                    shrinkWrap: true,
-                    scrollDirection: Axis.horizontal,
-                    itemBuilder: (context, index) {
-                      return Padding(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 8.0,
-                        ),
-                        child: Card(
-                          shadowColor: Color(0x44393e46),
-                          shape: RoundedRectangleBorder(
-                            borderRadius: BorderRadius.all(Radius.circular(20)),
-                          ),
-                          elevation: 5,
-                          child: Container(
-                            width: 300,
-                            child: Padding(
-                              padding: const EdgeInsets.all(8.0),
-                              child: Column(
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                  ListTile(
-                                    isThreeLine: true,
-                                    title: Text(
-                                      asyncSnapshot.data.documents[index].get(
-                                          'eventName'), // widget.scheduleList[index][],
-                                      style:
-                                          Theme.of(context).textTheme.headline6,
-                                    ),
-                                    subtitle: Row(
-                                      children: [
-                                        Icon(
-                                          Feather.map_pin,
-                                          size: 16.0,
-                                        ),
-                                        Text(
-                                          asyncSnapshot.data.documents[index]
-                                              .get('location'),
-                                          style: Theme.of(context)
-                                              .textTheme
-                                              .subtitle1,
-                                        ),
-                                      ],
-                                    ),
-                                    trailing: Column(
-                                      children: [
-                                        //time
-                                        Row(
-                                          mainAxisSize: MainAxisSize.min,
-                                          children: [
-                                            Text(DateFormat('E\ndd/MM\nkk:mm')
-                                                .format(asyncSnapshot
-                                                    .data.documents[index]
-                                                    .get('dateTime')
-                                                    .toDate())
-                                                .toString()),
-                                          ],
-                                        ),
-                                      ],
-                                    ),
-                                  ),
-                                ],
-                              ),
-                            ),
-                          ),
-                        ),
-                      );
-                    },
-                  )
-                : //if you have no events you will get this illustration
-                Container(
-                    child: Center(
-                      child: Image.asset("assets/events.png", height: 200),
-                    ),
-                  )
-            : Loader();
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-        height: MediaQuery.of(context).size.height / 5,
-        child: Stack(
-          children: <Widget>[feed()],
-        ));
-
-    //   ListView.builder(
-    //     scrollDirection: Axis.horizontal,
-    //     itemBuilder: (context, index) {
-    //       return Padding(
-    //         padding: const EdgeInsets.symmetric(
-    //           horizontal: 8.0,
-    //         ),
-    //         child: Card(
-    //           shadowColor: Color(0x44393e46),
-    //           shape: RoundedRectangleBorder(
-    //             borderRadius: BorderRadius.all(Radius.circular(20)),
-    //           ),
-    //           elevation: 5,
-    //           child: Container(
-    //             width: 300,
-    //             child: Padding(
-    //               padding: const EdgeInsets.all(8.0),
-    //               child: Column(
-    //                 crossAxisAlignment: CrossAxisAlignment.start,
-    //                 children: [
-    //                   ListTile(
-    //                     isThreeLine: true,
-    //                     title: Text(
-    //                       // widget.scheduleList[index][],
-    //                       style: Theme.of(context).textTheme.headline6,
-    //                     ),
-    //                     subtitle: Row(
-    //                       children: [
-    //                         Icon(
-    //                           Feather.map_pin,
-    //                           size: 16.0,
-    //                         ),
-    //                         Text(
-    //                           "Bahrain",
-    //                           style: Theme.of(context).textTheme.subtitle1,
-    //                         ),
-    //                       ],
-    //                     ),
-    //                     trailing: Column(
-    //                       children: [
-    //                         //time
-    //                         Row(
-    //                           mainAxisSize: MainAxisSize.min,
-    //                           children: [
-    //                             Text(DateTime.now().hour.toString()),
-    //                             Text(":"),
-    //                             Text(DateTime.now().minute.toString()),
-    //                           ],
-    //                         ),
-    //                         //day
-    //                         // Text(DateTime.now().weekday.toString()),
-    //                         if (DateTime.now().weekday == 1) Text("MON"),
-    //                         if (DateTime.now().weekday == 2) Text("TUE"),
-    //                         if (DateTime.now().weekday == 3) Text("WED"),
-    //                         if (DateTime.now().weekday == 4) Text("THU"),
-    //                         if (DateTime.now().weekday == 5) Text("FRI"),
-    //                         if (DateTime.now().weekday == 6) Text("SAT"),
-    //                         if (DateTime.now().weekday == 7) Text("SUN"),
-    //                       ],
-    //                     ),
-    //                   ),
-    //                 ],
-    //               ),
-    //             ),
-    //           ),
-    //         ),
-    //       );
-    //     },
-    //     itemCount: widget.scheduleList.length,
-    //   ),
-    // );
-  }
-}
 
 class ChatsTabs extends StatelessWidget {
   const ChatsTabs({
