+++ lib/view/schedule.dart
@@ -1,0 +1,139 @@
+import 'package:Runbhumi/services/EventService.dart';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_icons/flutter_icons.dart';
+import 'package:intl/intl.dart';
+import '../widget/widgets.dart';
+
+class Schedule extends StatefulWidget {
+  const Schedule({
+    Key key,
+  }) : super(key: key);
+
+  @override
+  _ScheduleState createState() => _ScheduleState();
+}
+
+class _ScheduleState extends State<Schedule> {
+  Stream currentFeed;
+  void initState() {
+    super.initState();
+    Firebase.initializeApp().whenComplete(() {
+      print("completed");
+      setState(() {});
+    });
+    getUserInfoEvents();
+  }
+
+  getUserInfoEvents() async {
+    EventService().getCurrentUserFeed().then((snapshots) {
+      setState(() {
+        currentFeed = snapshots;
+        print("we got the data + ${currentFeed.toString()} ");
+      });
+    });
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+  }
+
+  Widget feed() {
+    return StreamBuilder(
+      stream: currentFeed,
+      builder: (context, asyncSnapshot) {
+        print("Working");
+        return asyncSnapshot.hasData
+            ? asyncSnapshot.data.documents.length > 0
+                ? ListView.builder(
+                    itemCount: asyncSnapshot.data.documents.length,
+                    shrinkWrap: true,
+                    scrollDirection: Axis.horizontal,
+                    itemBuilder: (context, index) {
+                      return Padding(
+                        padding: const EdgeInsets.symmetric(
+                          horizontal: 8.0,
+                        ),
+                        child: Card(
+                          shadowColor: Color(0x44393e46),
+                          shape: RoundedRectangleBorder(
+                            borderRadius: BorderRadius.all(Radius.circular(20)),
+                          ),
+                          elevation: 5,
+                          child: Container(
+                            width: 300,
+                            child: Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Column(
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: [
+                                  ListTile(
+                                    isThreeLine: true,
+                                    title: Text(
+                                      asyncSnapshot.data.documents[index].get(
+                                          'eventName'), // widget.scheduleList[index][],
+                                      style:
+                                          Theme.of(context).textTheme.headline6,
+                                    ),
+                                    subtitle: Row(
+                                      children: [
+                                        Icon(
+                                          Feather.map_pin,
+                                          size: 16.0,
+                                        ),
+                                        Text(
+                                          asyncSnapshot.data.documents[index]
+                                              .get('location'),
+                                          style: Theme.of(context)
+                                              .textTheme
+                                              .subtitle1,
+                                        ),
+                                      ],
+                                    ),
+                                    trailing: Column(
+                                      children: [
+                                        //time
+                                        Row(
+                                          mainAxisSize: MainAxisSize.min,
+                                          children: [
+                                            Text(DateFormat('E\ndd/MM\nkk:mm')
+                                                .format(asyncSnapshot
+                                                    .data.documents[index]
+                                                    .get('dateTime')
+                                                    .toDate())
+                                                .toString()),
+                                          ],
+                                        ),
+                                      ],
+                                    ),
+                                  ),
+                                ],
+                              ),
+                            ),
+                          ),
+                        ),
+                      );
+                    },
+                  )
+                : //if you have no events you will get this illustration
+                Container(
+                    child: Center(
+                      child: Image.asset("assets/events.png", height: 200),
+                    ),
+                  )
+            : Loader();
+      },
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+        height: MediaQuery.of(context).size.height / 5,
+        child: Stack(
+          children: <Widget>[feed()],
+        ),
+      );
+  }
+}
