+++ lib/view/views.dart
@@ -8,6 +8,7 @@ export 'home.dart';
 export 'profilePage.dart';
 export 'networkPage.dart';
 export 'notificationsPage.dart';
+export 'schedule.dart';
 export 'drawer/drawer.dart';
 export 'drawer/moreInfo.dart';
 export 'drawer/editProfile.dart';
